feature updates -add list according to spellType -add key to each item on list
