+++ HPSpellBook/src/components/SpellTypes.js
@@ -1,35 +1,29 @@
 import API from "../../API";
 import { useEffect, useState } from "react";
-import { View, Text, ScrollView } from 'react-native';
-import uuid from 'react-native-uuid';
+import { View, Text } from 'react-native';
 
 const SpellTypes = () => {
 
-    const [spellList, setSpellList] = useState([]);
+    const [firstSpell, setFirstSpell] = useState([]);
 
-    const getSpellList = async () => {
+    const getSpellTypes = async () => {
         const { status, data } = await API.get('/Spells');
-        const spellList = data;
-
+        const firstSpell = data[0];
         if (status === 200) {
-            setSpellList(spellList);     
+            setFirstSpell(firstSpell);     
         }
     }
 
     useEffect(() => {
-        getSpellList();
+        getSpellTypes();
     }, []);
 
     return (
         <View>
-            <ScrollView >
-                {spellList.filter(spell => spell.type === 'Curse').map(filteredSpell => (
-                        <ScrollView key = {uuid.v4()}>
-                        <Text>{filteredSpell.name}</Text>
-                        </ScrollView>
-                ))  
-                }
-            </ScrollView>
+            <Text style={{fontWeight:'700'}}>Spell Types</Text>
+            <Text>Type: {firstSpell.type}</Text>
+            <Text>Name: {firstSpell.name}</Text>
+            <Text>Incantation: {firstSpell.incantation}</Text>
         </View>
     );
